backend/requirements.txt
requests
python-dotenv
flask
flask-cors
googlemaps
gunicorn